Move several helper functions to helpers/coverage.ts
