+++ src/components/ItemDetails.svelte
@@ -1,16 +1,10 @@
 <script lang="ts">
 import ItemBanner from './banners/ItemBanner.svelte'
 import MessageBanner from './banners/MessageBanner.svelte'
-import {
-  BillingPeriod,
-  isItemDraft,
-  ItemCoverageStatus,
-  itemIsApproved,
-  PolicyItem
-} from 'data/items'
+import { ItemCoverageStatus, PolicyItem } from 'data/items'
 import { getPolicyById, loadPolicy, policies, Policy, PolicyType } from 'data/policies'
-import { isBeforeMonthlyCutoff } from 'helpers/coverage'
-import { formatDate, startOfFutureMonth } from 'helpers/dates'
+import { getPremiumDescription, getRenewalDate } from 'helpers/coverage'
+import { formatDate } from 'helpers/dates'
 import { formatMoney } from 'helpers/money'
 import InfoBoxModal from './InfoBoxModal.svelte'
 import { formatDistanceToNow } from 'date-fns'
@@ -68,53 +62,6 @@ const getItemStatusText = (item: PolicyItem) => {
 
   return statusChangeStr + updatedAtStr
 }
-
-const getMonthlyRenewalDate = (item: PolicyItem): string => {
-  if (itemIsApproved(item)) {
-    return startOfFutureMonth(1)
-  }
-
-  if (isItemDraft(item)) {
-    if (isBeforeMonthlyCutoff()) {
-      return startOfFutureMonth(1)
-    } else {
-      return startOfFutureMonth(2)
-    }
-  }
-
-  return ''
-}
-
-const getRenewalDate = (item: PolicyItem | undefined): string => {
-  if (!item || !item.id) {
-    return ''
-  }
-
-  if (isMonthly(item)) {
-    return getMonthlyRenewalDate(item)
-  }
-
-  return getYearlyRenewalDate(item)
-}
-
-const getPremiumDescription = (item: PolicyItem | undefined): string => {
-  if (!item || !item.id) {
-    return ''
-  }
-
-  if (isMonthly(item)) {
-    return `${formatMoney(item.monthly_premium)} / month`
-  }
-
-  return `${formatMoney(item.annual_premium)} / year`
-}
-
-const getYearlyRenewalDate = (item: PolicyItem): string => {
-  const renewYear = new Date().getFullYear() + 1
-  return formatDate(`${renewYear}-01-01`)
-}
-
-const isMonthly = (item: PolicyItem) => item.billing_period === BillingPeriod.Monthly
 
 const toggleModal = (i: number) => (showInfoBox[i] = !showInfoBox[i])
 </script>
